proxy_server/requirements.txt
anyio==4.7.0
asgiref==3.8.1
certifi==2024.8.30
charset-normalizer==3.4.0
click==8.1.7
Django==5.1.4
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
h11==0.14.0
httptools==0.6.4
idna==3.10
PyJWT==2.10.1
python-dotenv==1.0.1
PyYAML==6.0.2
requests==2.32.3
sniffio==1.3.1
sqlparse==0.5.3
typing_extensions==4.12.2
urllib3==2.2.3
uvicorn==0.32.1
uvloop==0.21.0
watchfiles==1.0.0
websockets==14.1
